대전모드 커스터마이징 구현 / í,¤지 정리
