+++ src/main/java/client/ui/GameCustomizeUI.java
@@ -2,7 +2,7 @@ package client.ui;
 
 import client.ui.customize.CooperationCustomizeUI;
 import client.ui.customize.SpeedQuizCustomizeUI;
-import client.ui.customize.VersusCustomizeUI;
+import client.ui.customize.versusCustomizeUI.VersusCustomizeUI;
 import client.ui.icon.ArrowIcon;
 
 import javax.swing.*;
