+++ src/main/java/client/ui/RoomListUI.java
@@ -2,7 +2,7 @@ package client.ui;
 
 import client.ui.gamemode.CooperationUI;
 import client.ui.gamemode.SpeedQuizUI;
-import client.ui.gamemode.VersusUI;
+import client.ui.gamemode.versusUI.VersusUI;
 import client.ui.icon.ArrowIcon;
 import protocol.Message;
