--- src/main/java/client/ui/customize/VersusCustomizeUI.java
@@ -1,10 +1,0 @@
-package client.ui.customize;
-
-import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
-import java.net.Socket;
-
-public class VersusCustomizeUI {
-    public VersusCustomizeUI(Socket socket, ObjectOutputStream out, ObjectInputStream in, String loginUserId) {
-    }
-}
