+++ src/main/java/client/ui/customize/versusCustomizeUI/VersusCustomizeUI.java
@@ -1,0 +1,95 @@
+package client.ui.customize.versusCustomizeUI;
+
+import client.ui.GameCustomizeUI;
+import client.ui.icon.ArrowIcon;
+
+import javax.swing.*;
+import java.awt.*;
+import java.io.ObjectInputStream;
+import java.io.ObjectOutputStream;
+import java.net.Socket;
+
+public class VersusCustomizeUI {
+    private final Socket socket;
+    private final ObjectOutputStream out;
+    private final ObjectInputStream in;
+    private final String userId;
+    private String sendData;
+
+    public VersusCustomizeUI(Socket socket, ObjectOutputStream out, ObjectInputStream in, String userId) {
+        this.socket = socket;
+        this.out = out;
+        this.in = in;
+        this.userId = userId;
+
+        JFrame frame = new JFrame("Versus Customize");
+        frame.setSize(800, 600);
+        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+
+        JPanel panel = new JPanel();
+        panel.setLayout(null);
+        panel.setBackground(new Color(255, 247, 153));
+        frame.add(panel);
+
+        // 로고
+        JLabel logoLabel = new JLabel(new ImageIcon("./src/main/java/client/ui/nesquick_logo.png"));
+        logoLabel.setBounds(530, 10, 250, 100);
+        panel.add(logoLabel);
+
+        // 뒤로가기 버튼
+        JButton backButton = new JButton(new ArrowIcon(20, Color.BLACK));
+        backButton.setBounds(10, 10, 30, 30);
+        backButton.setBorderPainted(false);
+        backButton.setContentAreaFilled(false);
+        backButton.setFocusPainted(false);
+        panel.add(backButton);
+
+
+
+        // 퀴즈 제목
+        JLabel titleLabel = new JLabel("퀴즈 제목을 입력하세요");
+        titleLabel.setBounds(100, 200, 250, 30);
+        titleLabel.setFont(new Font("Malgun Gothic", Font.BOLD, 15));
+        panel.add(titleLabel);
+
+        JTextField titleField = new JTextField();
+        titleField.setBounds(300, 200, 300, 30);
+        panel.add(titleField);
+
+        // 문제 수
+        JLabel numLabel = new JLabel("문제 수를 선택하세요");
+        numLabel.setBounds(100, 300, 250, 30);
+        numLabel.setFont(new Font("Malgun Gothic", Font.BOLD, 15));
+        panel.add(numLabel);
+
+        String[] numbers = {"3", "4", "5"};
+        JComboBox<String> comboBox = new JComboBox<>(numbers);
+        comboBox.setBounds(300, 300, 70, 30);
+        panel.add(comboBox);
+
+        // 확인 버튼
+        JButton checkButton = new JButton("확인");
+        checkButton.setBounds(660, 470, 80, 30);
+        panel.add(checkButton);
+
+        frame.setVisible(true);
+
+
+
+        // 뒤로가기 버튼 동작
+        backButton.addActionListener(e -> {
+            frame.dispose();
+            new GameCustomizeUI(socket, out, in, userId);
+        });
+
+        // 확인 버튼 동작
+        checkButton.addActionListener(e -> {
+            String title = titleField.getText();
+            String selectedValue = (String) comboBox.getSelectedItem();
+            sendData = title+"\n"+selectedValue+"\n";
+            frame.dispose();
+            new VersusCustomizeGetQuizUI(socket, out, in, userId, Integer.parseInt(selectedValue),sendData);
+        });
+
+    }
+}
